requirements.txt
discord.py==2.5.0
SQLAlchemy==2.0.36
psycopg[binary]==3.3.2
python-dotenv==1.0.0